# internal/api/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_ash-repwiki_ash_internal_spacepolicy.EffectivePolicy:
    properties:
      citationMode:
        type: string
      multiSign:
        type: boolean
      reviewSlaHours:
        type: integer
      sources:
        items:
          type: string
        type: array
      spaceId:
        type: string
      spaceKind:
        type: string
    type: object
  github_com_ash-repwiki_ash_internal_spacepolicy.Pack:
    properties:
      bodyJson:
        type: string
      citationMode:
        type: string
      multiSign:
        type: boolean
      reviewSlaHours:
        type: integer
      spaceId:
        type: string
      updatedAt:
        type: integer
    type: object
  github_com_ash-repwiki_ash_internal_spacepolicy.PutPackRequest:
    properties:
      bodyJson:
        type: string
      citationMode:
        type: string
      multiSign:
        type: boolean
      reviewSlaHours:
        type: integer
    type: object
...
  internal_api.spacePolicyResponse:
    properties:
      effective:
        $ref: '#/definitions/github_com_ash-repwiki_ash_internal_spacepolicy.EffectivePolicy'
      pack:
        $ref: '#/definitions/github_com_ash-repwiki_ash_internal_spacepolicy.Pack'
    type: object
...
host: localhost:8080
info:
  contact:
    name: ASH
  description: Delivery-oriented coding robot worker API (M0). See docs/appendices/G-OpenAPI-端点清单(M0).md.
  license:
    name: Proprietary
  title: ASH Worker API
  version: "0.1"
paths:
...
  /api/v1/spaces/{spaceId}/policy:
    get:
      parameters:
      - description: space id
        in: path
        name: spaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.spacePolicyResponse'
      summary: Get space policy pack and effective merge
      tags:
      - spaces
    put:
      consumes:
      - application/json
      parameters:
      - description: space id
        in: path
        name: spaceId
        required: true
        type: string
      - description: pack
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ash-repwiki_ash_internal_spacepolicy.PutPackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.spacePolicyResponse'
      summary: Upsert space policy pack
      tags:
      - spaces
...
swagger: "2.0"
...